requirements-test.txt
--requirement requirements.txt
# With the release of version 2.10, Ansible finally correctly
# identifies Kali Linux as being the Kali distribution of the Debian
# OS family.  This simplifies a lot of things for roles that support
# Kali Linux, so it makes sense to force the installation of Ansible
# 2.10 or newer.
ansible>=2.10,<7
ansible-lint>=5,<6
flake8
molecule[docker]
pre-commit
pytest-testinfra
yamllint
requests